requirements.txt
jinja2
PyYaml
#python-gnupg==0.4.3
cookiecutter==1.6.0
#cryptography
